==> 1_get_data.ipynb <==
# Weekly Counts of Deaths by Jurisdiction and Age

* Download the CDC data from https://data.cdc.gov/NCHS/Weekly-Counts-of-Deaths-by-Jurisdiction-and-Age/y5bj-9g5w
* Download the NYT COVID-19 Deaths from https://github.com/nytimes/covid-19-data
* Download population data from https://www2.census.gov/programs-surveys/popest/datasets/2010-2020/state/asrh/
* Merge all into `data/us_excess_deaths.rds`

library(tidyverse)

options(readr.show_col_types = F)

download.file("https://data.cdc.gov/api/views/y5bj-9g5w/rows.csv?accessType=DOWNLOAD",
             destfile = "data/Weekly_Counts_of_Deaths_by_Jurisdiction_and_Age.csv")

# Make one age bin of all age groups <65

df <- read_csv('data/Weekly_Counts_of_Deaths_by_Jurisdiction_and_Age.csv',
              col_types = "ccciicicccc")%>%
  filter(Type != 'Predicted (weighted)')%>%
  select(Jurisdiction, Year, Week, `Number of Deaths`, `Age Group`)%>%
  mutate (age_lower = parse_number(gsub("Under 25 years", "0", `Age Group`)))%>%
  mutate(`Age Group` = case_when(age_lower<65 ~ "<65",
                                 age_lower<75 ~ "65-74",
                                 age_lower<85 ~ "75-84",
                                            T ~ "85+"))%>%
  group_by(Jurisdiction, Year, Week, `Age Group`)%>%
  summarise(`Number of Deaths` = sum(ifelse(is.na(`Number of Deaths`), 0, `Number of Deaths`)))

head(df)

# Merge 'New York' and 'New York City'
In the CDC data 'New York' means 'New York *excluding* New York City'

df2 <-
  df%>%
  mutate(Jurisdiction = gsub("New York City", "New York", Jurisdiction))%>%
  group_by(Jurisdiction, Year, `Age Group`, Week)%>%
  summarise(`Number of Deaths` = sum(`Number of Deaths`))

write_csv(df2, "data/tidy_deaths_age_state_CDC.csv")

head(df2)

# Download population data
See https://www2.census.gov/programs-surveys/popest/technical-documentation/file-layouts/2010-2020/sc-est2020-agesex-civ.pdf
for a description

download.file("https://www2.census.gov/programs-surveys/popest/datasets/2010-2020/state/asrh/SC-EST2020-AGESEX-CIV.csv",
             destfile = "data/Population.csv")

df_pop <- read_csv('data/Population.csv')
head(df_pop)

df_pop2 <-  
  df_pop%>%
  filter(SEX==0 & AGE != 999)%>%
  mutate(`Age Group` = case_when(AGE<65 ~ "<65",
                                 AGE<75 ~ "65-74",
                                 AGE<85 ~ "75-84",
                                      T ~ "85+"))%>%
  group_by(NAME, `Age Group`)%>%
  summarise(Population = sum(POPEST2020_CIV))%>%
  rename(Jurisdiction = NAME)

write_csv(df_pop2, "data/tidy_pop_age_state.csv")

head(df_pop2)

# Excess Deaths Associated with COVID-19
Download the CDC data from https://www.cdc.gov/nchs/nvss/vsrr/covid19/excess_deaths.htm

download.file("https://data.cdc.gov/api/views/xkkf-xrst/rows.csv?accessType=DOWNLOAD",
             destfile = "data/Excess_Deaths_Associated_with_COVID-19.csv")
head(df)

df <- 
  read_csv("data/Excess_Deaths_Associated_with_COVID-19.csv")%>%
  filter(Type=="Unweighted")%>%
  select(State, `Week Ending Date`, `Average Expected Count`, `Observed Number`)%>%
  mutate(State = gsub("New York City", "New York", State))%>%
  group_by(`Week Ending Date`, State)%>%
  summarise(`Average Expected Count` = sum(`Average Expected Count`),
           `Observed Number` = sum(`Observed Number`))

write_csv(df, "data/tidy_deaths_state_CDC.csv")

head(df)

# Download COVID-19 deaths from the New York Times

df_covid <- 
  read_csv("https://github.com/nytimes/covid-19-data/raw/master/us-states.csv")%>%
  group_by(state)%>%
  arrange(date)%>%
  mutate(deaths = deaths-lag(deaths, 1, default = 0))%>%
  mutate(`Covid Deaths` = zoo::rollmean(deaths, k=21, na.pad = T))%>%
  rename(`Week Ending Date` = date,
         State = state)%>%
  select(`Week Ending Date`, State, `Covid Deaths`)%>%
  na.omit()%>%
  ungroup()

# Sum up all states for the 'United States' estimate
df_covid <- 
rbind.data.frame(df_covid,
  df_covid%>%
    group_by(`Week Ending Date`)%>%
    summarise(`Covid Deaths` = sum(`Covid Deaths`))%>%
    ungroup()%>%
    mutate(State = "United States")%>%
    select(`Week Ending Date`, `State`, `Covid Deaths`)
  )
tail(df_covid)

df2 <-
  df%>%
  select(State, `Week Ending Date`, `Average Expected Count`, `Observed Number`)%>%
  merge(df_covid, all = T)%>%
  drop_na(`Observed Number`)%>%
  merge(
    read_csv("data/tidy_pop_age_state.csv", show_col_types = FALSE)%>%
      group_by(Jurisdiction)%>%
      summarise(Population = sum(Population))%>%
      rename(State = Jurisdiction)
  )

write_rds(df2, "data/us_excess_deaths.rds")

head(df2)
