==> 3_merge_data.ipynb <==
# Merge the data
Create the `data/us_excess_per_age_group.rds` data frame.

library(tidyverse)

df <-
  rbind(read_csv("data/tidy_deaths_age_state_CDC.csv", show_col_types = FALSE)%>%
          mutate(Type = "Observed"), 
        read_csv("data/tidy_predicted_deaths_age_state_CDC.csv", show_col_types = FALSE)%>%
          mutate(Type = "Predicted"))%>%
  mutate(Date = ISOweek::ISOweek2date(paste0(Year, "-W", str_pad(Week, 2, pad = "0"), "-7")))

head(df)

# Read population data
For the age normalization we need the following fields:
* __Population__: The population of the age group for the state
* __Population_all_ages__: The population of the entire state
* __Population_age_group_nation__: The population of the age group for the entire nation

df_pop1 <- read_csv("data/tidy_pop_age_state.csv", show_col_types = FALSE)

df_pop2 <- 
  df_pop1%>%
  group_by(Jurisdiction)%>%
  summarise(Population_all_ages = sum(Population))

df_pop3 <- 
  df_pop1%>%
  filter(Jurisdiction=="United States")%>%
  rename(Population_age_group_nation = Population)%>%
  select(-Jurisdiction)

df_pop <- 
  df_pop1%>%
  merge(df_pop2)%>%
  merge(df_pop3)

df2 <- merge(df, df_pop)
write_rds(df2, "data/us_excess_deaths_per_age_group.rds")
head(df2)
